app/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cnn-finetune==0.6.0
cycler==0.10.0
Flask==1.1.1
Flask-Dropzone==1.5.4
Flask-Uploads==0.2.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
munch==2.5.0
numpy==1.17.4
Pillow==6.2.1
pretrainedmodels==0.7.4
pyparsing==2.4.5
python-dateutil==2.8.1
requests==2.22.0
scipy==1.3.2
six==1.13.0
torch==1.3.1
torchvision==0.4.2
tqdm==4.38.0
urllib3==1.25.7
Werkzeug==0.16.0
